background-remover/requirements.txt
git+https://github.com/facebookresearch/segment-anything.git
torch 
torchvision 
opencv-python 
numpy
matplotlib
Pillow
streamlit
streamlit-image-coordinates
streamlit-dimensions
